kasipOdaq/kasipodaq-backend-master/requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.8
luhn==0.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openpyxl==3.0.6
openpyexcel==2.5.14
packaging==20.3
paramiko==2.7.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
qrcode==6.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.17.8
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
